Move helpers into component helpers
