+++ purview.cabal
@@ -44,6 +44,7 @@ library
       CleanTree
       CollectInitials
       Component
+      ComponentHelpers
       Diffing
       EventHandling
       EventLoop
@@ -85,6 +86,7 @@ test-suite purview-test
       CleanTree
       CollectInitials
       Component
+      ComponentHelpers
       Diffing
       EventHandling
       EventLoop
